Publish after testing
Added code to check if a given pdf is actually an annex form and to ignore pdfs that don't contain text from the form title. Also added handling for when no valid annex data was found and added content to the box formatting errors to declare the file name when a formatting error is found.